Fix Windows cross-compile 32-bit truncation bugs in transaction handling
Critical fixes for Windows MinGW cross-compilation where uint64_t values
were being truncated to 32-bit, causing invalid transactions.

Issues fixed:
1. Integer arithmetic overflow in manual parsing loops
   - Changed `* 10` to `* 10ULL` to force 64-bit multiplication
   - Prevents overflow when parsing datoshi amounts

2. Union member aliasing bug in GET_256_FROM_64 macro
   - Changed writes from `._lo.a` to `.lo.lo` to match reads
   - Ensures consistent union member access across platforms

3. Printf format specifiers truncating 64-bit values
   - Changed all `%lu` to `%llu` with (unsigned long long) casts
   - Affects debug output and JSON conversion

4. JSON conversion truncating transaction amounts
   - cellframe_json_minimal.c: sprintf("%lu") -> sprintf("%llu")
   - Was causing "Sign verification failed" due to wrong amounts in JSON

5. GUI rewrites
   - Replaced QDoubleSpinBox with QLineEdit to avoid float precision loss
   - Completely rewrote GUI send logic copying exact flow from CLI

All transactions now properly handle 64-bit datoshi values on Windows.
Tested and verified on both Linux native and Windows cross-compiled builds.

+++ cellframe_json_minimal.c
@@ -93,7 +93,7 @@ void cellframe_uint256_to_str(const uint256_t *value, char *str_out) {
 
     // For amounts < 2^64, value is in lo.lo (bytes 16-23)
     // SDK: uint256_t.lo.lo contains the datoshi amount
-    sprintf(str_out, "%lu", value->lo.lo);
+    sprintf(str_out, "%llu", (unsigned long long)value->lo.lo);
 }
 
 // ============================================================================
